# Tema 7/LaravelWebServices/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Productos",
        "description": "API de productos",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login",
                "description": "Login del usuario",
                "operationId": "login",
                "requestBody": {
                    "description": "Datos del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "prueba@prueba.es"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login correcto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "description": "Retorna una lista de todos los productos almacenados en la base de datos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron productos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un producto",
                "description": "Crea un nuevo producto en el sistema. Requiere autenticación JWT.",
                "operationId": "store",
                "requestBody": {
                    "description": "Datos del producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "stock"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Producto 1"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del producto"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto",
                "description": "Obtiene un producto por su id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Categoria": {
                "title": "Categoria",
                "description": "Modelo de una categoría",
                "properties": {
                    "id": {
                        "description": "ID único de la categoría",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la categoría",
                        "type": "string",
                        "example": "Categoria 1"
                    },
                    "productos": {
                        "description": "Lista de productos asociados a la categoría",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Producto"
                        }
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "description": "Modelo de un producto",
                "properties": {
                    "id": {
                        "description": "ID del producto",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Laptop"
                    },
                    "precio": {
                        "description": "Precio del producto",
                        "type": "number",
                        "format": "float",
                        "example": 899.99
                    },
                    "stock": {
                        "description": "Cantidad en stock",
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Producto"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use un token JWT para autenticarse",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "login",
            "description": "login"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}